+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/types/internal/infer/TypeInferenceLogger.java
@@ -29,6 +29,7 @@ import net.sourceforge.pmd.lang.java.types.JTypeMirror;
 import net.sourceforge.pmd.lang.java.types.TypePrettyPrint;
 import net.sourceforge.pmd.lang.java.types.TypePrettyPrint.TypePrettyPrinter;
 import net.sourceforge.pmd.lang.java.types.internal.infer.ExprMirror.CtorInvocationMirror;
+import net.sourceforge.pmd.lang.java.types.internal.infer.ExprMirror.FunctionalExprMirror;
 import net.sourceforge.pmd.lang.java.types.internal.infer.ExprMirror.InvocationMirror.MethodCtDecl;
 import net.sourceforge.pmd.lang.java.types.internal.infer.InferenceVar.BoundKind;
 import net.sourceforge.pmd.util.StringUtil;
@@ -80,6 +81,8 @@ public interface TypeInferenceLogger {
     default void skipArgAsNonPertinent(int i, ExprMirror expr) { }
 
     default void functionalExprNeedsInvocationCtx(JTypeMirror targetT, ExprMirror expr) { }
+
+    default void functionalExprHasUnresolvedTargetType(JTypeMirror targetT, FunctionalExprMirror expr) { }
 
     default void endArg() { }
 
@@ -304,6 +307,12 @@ public interface TypeInferenceLogger {
         }
 
         @Override
+        public void functionalExprHasUnresolvedTargetType(JTypeMirror targetT, FunctionalExprMirror expr) {
+            println("[WARNING] Target type for functional expression is unresolved: " + targetT);
+            println("Will treat the expression as matching (this may cause future mistakes)");
+        }
+
+        @Override
         public void ambiguityError(MethodCallSite site, @Nullable MethodCtDecl selected, List<MethodCtDecl> methods) {
             println("");
             printExpr(site.getExpr());
@@ -450,6 +459,7 @@ public interface TypeInferenceLogger {
             println("Target type is not a functional interface yet: " + targetT);
             println("Will wait for invocation phase before discarding.");
         }
+
 
         @Override
         public void endArgsChecks() {